+++ PremiumScraps/CustomEffects/LegendaryStick.cs
@@ -1,9 +1,9 @@
 ﻿using GameNetcodeStuff;
-using LethalNetworkAPI;
 using PremiumScraps.Utils;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using Unity.Netcode;
 using UnityEngine;
 
 namespace PremiumScraps.CustomEffects
@@ -12,34 +12,16 @@ namespace PremiumScraps.CustomEffects
     {
         public int knockbackPowerMin = 1;
         public int knockbackPowerMax = 10;
-        public readonly bool specialEffectEnabled = true;
         public int weaponHitForce = 1;
         public bool reelingUp;
         public bool isHoldingButton;
-        private Coroutine reelingUpCoroutine;
+        private Coroutine? reelingUpCoroutine;
         private RaycastHit[] objectsHitByWeapon;
         private List<RaycastHit> objectsHitByWeaponList = new List<RaycastHit>();
         private PlayerControllerB previousPlayerHeldBy;
         private readonly int weaponMask = 11012424;
-        public LethalClientMessage<PosId> network;
-        public LethalClientMessage<PosId> networkAudio;
-        public LegendaryStick()
-        {
-            network = new LethalClientMessage<PosId>(identifier: "premiumscrapsStickID");
-            networkAudio = new LethalClientMessage<PosId>(identifier: "premiumscrapsStickAudioID");
-            network.OnReceivedFromClient += KnockbackNetwork;
-            networkAudio.OnReceivedFromClient += InvokeAudioNetwork;
-        }
 
-        private void KnockbackNetwork(PosId info, ulong clientId)
-        {
-            Effects.Knockback(info.position, 1, 0, info.Id);
-        }
-
-        private void InvokeAudioNetwork(PosId posId, ulong clientId)
-        {
-            Effects.Audio(posId.Id, posId.position, 2.5f);
-        }
+        public LegendaryStick() { }
 
         public override void ItemActivate(bool used, bool buttonDown = true)
         {
@@ -72,7 +54,7 @@ namespace PremiumScraps.CustomEffects
             playerHeldBy.twoHanded = true;
             playerHeldBy.playerBodyAnimator.ResetTrigger("stickHit");
             playerHeldBy.playerBodyAnimator.SetBool("reelingUp", true);
-            networkAudio.SendAllClients(new PosId(3, playerHeldBy.transform.position));
+            AudioServerRpc(3, playerHeldBy.transform.position, 2.5f);
             yield return new WaitForSeconds(0.35f);
             yield return new WaitUntil(() => !isHoldingButton || !isHeld);
             SwingWeapon(!isHeld);
@@ -88,7 +70,7 @@ namespace PremiumScraps.CustomEffects
             previousPlayerHeldBy.playerBodyAnimator.SetBool("reelingUp", value: false);
             if (!cancel)
             {
-                networkAudio.SendAllClients(new PosId(4, previousPlayerHeldBy.transform.position));
+                AudioServerRpc(4, previousPlayerHeldBy.transform.position, 2.5f);
                 previousPlayerHeldBy.UpdateSpecialAnimationValue(specialAnimation: true, (short)previousPlayerHeldBy.transform.localEulerAngles.y, 0.4f);
             }
         }
@@ -105,14 +87,14 @@ namespace PremiumScraps.CustomEffects
             if (!cancel)
             {
                 previousPlayerHeldBy.twoHanded = false;
-                Debug.DrawRay(previousPlayerHeldBy.gameplayCamera.transform.position + previousPlayerHeldBy.gameplayCamera.transform.right * -0.35f, previousPlayerHeldBy.gameplayCamera.transform.forward * 1.85f, Color.blue, 5f);
-                objectsHitByWeapon = Physics.SphereCastAll(previousPlayerHeldBy.gameplayCamera.transform.position + previousPlayerHeldBy.gameplayCamera.transform.right * -0.35f, 0.75f, previousPlayerHeldBy.gameplayCamera.transform.forward, 1.85f, weaponMask, QueryTriggerInteraction.Collide);
+                Debug.DrawRay(previousPlayerHeldBy.gameplayCamera.transform.position + previousPlayerHeldBy.gameplayCamera.transform.right * -0.35f,
+                    previousPlayerHeldBy.gameplayCamera.transform.forward * 1.85f, Color.blue, 5f);
+                objectsHitByWeapon = Physics.SphereCastAll(previousPlayerHeldBy.gameplayCamera.transform.position + previousPlayerHeldBy.gameplayCamera.transform.right * -0.35f,
+                    0.75f, previousPlayerHeldBy.gameplayCamera.transform.forward, 1.85f, weaponMask, QueryTriggerInteraction.Collide);
                 objectsHitByWeaponList = objectsHitByWeapon.OrderBy((x) => x.distance).ToList();
                 Vector3 start = previousPlayerHeldBy.gameplayCamera.transform.position;
                 for (int i = 0; i < objectsHitByWeaponList.Count; i++)
                 {
-                    IHittable component;
-                    RaycastHit hitInfo;
                     if (objectsHitByWeaponList[i].transform.gameObject.layer == 8 || objectsHitByWeaponList[i].transform.gameObject.layer == 11)
                     {
                         start = objectsHitByWeaponList[i].point + objectsHitByWeaponList[i].normal * 0.01f;
@@ -128,23 +110,18 @@ namespace PremiumScraps.CustomEffects
                             }
                         }
                     }
-                    else if (objectsHitByWeaponList[i].transform.TryGetComponent(out component) && !(objectsHitByWeaponList[i].transform == previousPlayerHeldBy.transform) && (objectsHitByWeaponList[i].point == Vector3.zero || !Physics.Linecast(start, objectsHitByWeaponList[i].point, out hitInfo, StartOfRound.Instance.collidersAndRoomMaskAndDefault))
-                        && !StartOfRound.Instance.inShipPhase)
+                    else if (objectsHitByWeaponList[i].transform.TryGetComponent(out IHittable component) && !(objectsHitByWeaponList[i].transform == previousPlayerHeldBy.transform)
+                        && (objectsHitByWeaponList[i].point == Vector3.zero || !Physics.Linecast(start, objectsHitByWeaponList[i].point, out RaycastHit hitInfo, StartOfRound.Instance.collidersAndRoomMaskAndDefault)))
                     {
                         flag = true;
-
-                        if (specialEffectEnabled && component.GetType() == typeof(PlayerControllerB))
+                        if (component.GetType() == typeof(PlayerControllerB))
                         {
                             if (!((PlayerControllerB)component).inSpecialInteractAnimation)
                             {
                                 int power = Random.Range(knockbackPowerMin, knockbackPowerMax);
-                                network.SendAllClients(new PosId(power, previousPlayerHeldBy.transform.position), false);
+                                KnockbackServerRpc(previousPlayerHeldBy.transform.position, power);
                             }
                         }
-                        //else
-                        //{
-                        //component.Hit(weaponHitForce, forward, previousPlayerHeldBy, true);
-                        //}
                     }
                 }
             }
@@ -152,8 +129,34 @@ namespace PremiumScraps.CustomEffects
             {
                 FindObjectOfType<RoundManager>().PlayAudibleNoise(transform.position, 17f, 0.8f);
                 playerHeldBy.playerBodyAnimator.SetTrigger("stickHit");
-                networkAudio.SendAllClients(new PosId(5, playerHeldBy.transform.position));
+                AudioServerRpc(5, playerHeldBy.transform.position, 2.5f);
             }
+        }
+
+        [ServerRpc(RequireOwnership = false)]
+        private void KnockbackServerRpc(Vector3 position, int power)
+        {
+            KnockbackClientRpc(position, power);
+        }
+
+        [ClientRpc]
+        private void KnockbackClientRpc(Vector3 position, int power)
+        {
+            if (playerHeldBy != null && GameNetworkManager.Instance.localPlayerController.playerClientId == playerHeldBy.playerClientId)
+                return;
+            Effects.Knockback(position, 1, 0, power);
+        }
+
+        [ServerRpc(RequireOwnership = false)]
+        private void AudioServerRpc(int audioID, Vector3 position, float volume)
+        {
+            AudioClientRpc(audioID, position, volume);
+        }
+
+        [ClientRpc]
+        private void AudioClientRpc(int audioID, Vector3 position, float volume)
+        {
+            Effects.Audio(audioID, position, volume);
         }
     }
 }