+++ PremiumScraps/Plugin.cs
@@ -11,7 +11,6 @@ using UnityEngine;
 
 namespace PremiumScraps
 {
-    [BepInDependency("LethalNetworkAPI")]
     [BepInPlugin(GUID, NAME, VERSION)]
     public class Plugin : BaseUnityPlugin
     {
@@ -38,6 +37,8 @@ namespace PremiumScraps
                 case 6: script = item.spawnPrefab.AddComponent<JobDark>(); break;
                 case 7: script = item.spawnPrefab.AddComponent<SpanishDrink>(); break;
                 case 8: script = item.spawnPrefab.AddComponent<TalkingBall>(); SetupScript.Copy((SoccerBallProp)script, item); break;
+                case 9: script = item.spawnPrefab.AddComponent<HarryDoll>(); break;
+                case 10: script = item.spawnPrefab.AddComponent<BadStatue>(); break;
                 default: return;
             }
             script.grabbable = true;
@@ -87,7 +88,7 @@ namespace PremiumScraps
                 new Scrap("AinzOoalGown/AinzOoalGownItem.asset", 5),
                 new Scrap("HelmDomination/HelmDominationItem.asset", 11),
                 new Scrap("TheKing/TheKingItem.asset", 13),
-                new Scrap("HarryMason/HarryMasonItem.asset", 10),
+                new Scrap("HarryMason/HarryMasonItem.asset", 10, 9),
                 new Scrap("Cristal/CristalItem.asset", 9),
                 new Scrap("PuppyShark/PuppySharkItem.asset", 10),
                 new Scrap("Rupee/RupeeItem.asset", 15),
@@ -97,7 +98,7 @@ namespace PremiumScraps
                 new Scrap("Spoon/SpoonItem.asset", 12),
                 new Scrap("Crouton/CroutonItem.asset", 6),
                 new Scrap("AirHornCustom/AirHornCustomItem.asset", 11, 1),
-                new Scrap("Balan/BalanItem.asset", 10),
+                new Scrap("Balan/BalanItem.asset", 10, 10),
                 new Scrap("CustomFace/CustomFaceItem.asset", 8, 2),
                 new Scrap("Scroll/ScrollItem.asset", 7, 3),
                 new Scrap("Stick/StickItem.asset", 9, 4),